% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{adicionar_paginacao}
\alias{adicionar_paginacao}
\title{Adiciona paginacao a uma url}
\usage{
adicionar_paginacao(url)
}
\arguments{
\item{url}{url da primeira pagina da busca gerada por
\code{\link{formar_url}}}
}
\value{
um vetor de strings com urls de cada pagina da busca
}
\description{
Adiciona paginacao a uma url
}
