% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{formar_url}
\alias{formar_url}
\title{Formar url com termo e periodo da busca}
\usage{
formar_url(termo, inicio, fim)
}
\arguments{
\item{termo}{termo buscado}

\item{inicio}{inicio do periodo da busca}

\item{fim}{fim do periodo da busca}
}
\value{
uma string com url da busca
}
\description{
Formar url com termo e periodo da busca
}
