% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{gentilmente}
\alias{gentilmente}
\title{Roda gentilmente}
\usage{
gentilmente(.f, mean = 1)
}
\arguments{
\item{.f}{funcao a rodar}

\item{mean}{tempo medio de espera entre uma execucao e outra}
}
\value{
uma funcao que executa \code{.f} depois de um intervalo
aletorio com media \code{mean}
}
\description{
Roda gentilmente
}
\examples{
media_gentil <- folhar2:::gentilmente(mean)
system.time(media_gentil(1:9))
}
