% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{listar_textos}
\alias{listar_textos}
\title{Lista textos de uma colunista}
\usage{
listar_textos(colunista)
}
\arguments{
\item{colunista}{Nome da colunista}
}
\value{
uma \code{\link{tibble}} com dados dos textos da colunista
}
\description{
Lista textos de uma colunista
}
\examples{
listar_textos("Denise Fraga")
}
