% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_busca}
\alias{parse_busca}
\title{Faz parse da busca}
\usage{
parse_busca(link)
}
\arguments{
\item{link}{url de uma pagina da busca}
}
\value{
uma \code{\link{tibble}} com informacoes dos resultados da busca
}
\description{
Faz parse da busca
}
