% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{folha_buscar}
\alias{folha_buscar}
\title{Buscar na folha}
\usage{
folha_buscar(termo, inicio = NULL, fim = NULL, media_espera = 1)
}
\arguments{
\item{termo}{termo buscado}

\item{inicio}{inicio do periodo da busca}

\item{fim}{fim do periodo da busca}

\item{media_espera}{media do tempo de espera entre requisicoes}
}
\value{
uma \code{\link{tibble}} com informacoes dos resultados da busca
}
\description{
Buscar na folha
}
\examples{
\dontrun{
folha_buscar("recessao", "01/01/2008", "01/07/2008")
}

}
